restructure code as a module
