+++ classer/embedders/base.py
