+++ classer/embedders/hash.py
