+++ classer/models/hash_nb.py
